integrations/CrewAI-SidClaw/requirements.txt
crewai>=0.130.0
sidclaw>=0.1.2
python-dotenv>=1.0.0
